+++ src/core/alghoritms/filter.h
@@ -6,6 +6,6 @@
 enum filters { BOX_FILTER, MED_FILTER, GAUSSIAN_FILTER, UNSHARP_FILTER };
 
 void filter_image(Matrix *matrix, int max_color_value, int filter_size,
-                  enum filters filter, float sigma);
+                  enum filters filter, double sigma);
 
 #endif // FILTER_H
