+++ src/core/math/vector.c
@@ -3,8 +3,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-float mult_vectors(Vector *vec1, Vector *vec2) {
-    float result = 0;
+double mult_vectors(Vector *vec1, Vector *vec2) {
+    double result = 0;
     for (size_t i = 0; i < vec1->size; i++)
         result += vec1->data[i] * vec2->data[i];
     return result;
@@ -13,20 +13,20 @@ float mult_vectors(Vector *vec1, Vector *vec2) {
 Vector *create_vector(size_t size) {
     Vector *vec = (Vector *)malloc(sizeof(Vector));
     vec->size = size;
-    vec->data = (float *)malloc(sizeof(float) * size);
+    vec->data = (double *)malloc(sizeof(double) * size);
     return vec;
 }
 
-float sum_vector(Vector *vec) {
-    float result = 0;
+double sum_vector(Vector *vec) {
+    double result = 0;
     for (size_t i = 0; i < vec->size; i++)
         result += vec->data[i];
     return result;
 }
 
 int _comp(const void *elem1, const void *elem2) {
-    int f = *((float *)elem1);
-    int s = *((float *)elem2);
+    int f = *((double *)elem1);
+    int s = *((double *)elem2);
     if (f > s)
         return 1;
     if (f < s)
@@ -35,7 +35,7 @@ int _comp(const void *elem1, const void *elem2) {
 }
 
 void sort_vector(Vector *vec) {
-    qsort(vec->data, vec->size, sizeof(float), _comp);
+    qsort(vec->data, vec->size, sizeof(double), _comp);
 }
 
 void free_vector(Vector *vec) {
