+++ src/core/math/vector.h
@@ -6,14 +6,14 @@
 
 typedef struct {
     size_t size;
-    float *data;
+    double *data;
 } Vector;
 
 Vector *create_vector(size_t size);
 void free_vector(Vector *vec);
 
-float mult_vectors(Vector *vec1, Vector *vec2);
-float sum_vector(Vector *vec);
+double mult_vectors(Vector *vec1, Vector *vec2);
+double sum_vector(Vector *vec);
 void sort_vector(Vector *vec);
 
 #endif // VECTOR_H
